Fix switch statement fall-through warnings with gcc 7.2. bugref:8192: gcc warnings
gcc 7.1 and later add a switch statement fall-through warning level, which
-Wall sets to level 3.  At this level, fall-throughs have to have at least
a comment following particular requirements (see gcc manual).  This change
fixes a few places in the code to meet these requirements.  Currently this
warning prevents building with kObjCache enabled, as the compiler checks
the comments in the source, which are stripped out by the object cache.

git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@68994 cfe28804-0f27-0410-a406-dd0f0b0b656f